Clear code-health sweep quick wins: dead file, cycle imports, dupe helpers
Delete resolveModel.ts — knip --production's only unused-file hit, no
importers outside its own test, and the docblock named routes that
don't actually call it.

Break 5 of the 6 dependency cycles skott was reporting, all barrel
self-imports: three lib/utils files pulled a formatter back through
their own index, ai-testing.types imported World/Character/
NarrativeContext through types/index, and ThemeProvider imported its
own color-scheme constants back through theme/index. Point each at
the file that actually defines the thing (extracted colorScheme.ts
for the theme case, since those constants lived in index.ts itself).
Cycle count drops from 6 to 1; lowered .skott-baseline.json to match.
Re-baselined dependency-cruiser too, since 5 suppressed violations no
longer reproduce (17 -> 12).

Dedupe the isRecord/asTrimmedString/asArray triplet that landed
byte-identical in worldCostExtraction.ts and worldThreadExtraction.ts
this week, plus a third spelling of the trim helper in
structuredLoreExtractor.ts, into typeGuards/jsonGuards.ts alongside
the existing stringGuards/worldGuards modules.

Narraitor#1951 findings 4, 5, 7.
