+++ .dependency-cruiser-known-violations.json
@@ -187,138 +187,5 @@
         ]
       }
     ]
-  },
-  {
-    "type": "cycle",
-    "from": "src/lib/theme/index.ts",
-    "to": "src/lib/theme/ThemeProvider.tsx",
-    "rule": {
-      "severity": "warn",
-      "name": "no-circular"
-    },
-    "cycle": [
-      {
-        "name": "src/lib/theme/ThemeProvider.tsx",
-        "dependencyTypes": [
-          "local",
-          "export"
-        ]
-      },
-      {
-        "name": "src/lib/theme/index.ts",
-        "dependencyTypes": [
-          "local",
-          "import"
-        ]
-      }
-    ]
-  },
-  {
-    "type": "cycle",
-    "from": "src/lib/utils/enumHelpers.ts",
-    "to": "src/lib/utils/index.ts",
-    "rule": {
-      "severity": "warn",
-      "name": "no-circular"
-    },
-    "cycle": [
-      {
-        "name": "src/lib/utils/index.ts",
-        "dependencyTypes": [
-          "aliased",
-          "aliased-tsconfig",
-          "aliased-tsconfig-paths",
-          "local",
-          "import"
-        ]
-      },
-      {
-        "name": "src/lib/utils/enumHelpers.ts",
-        "dependencyTypes": [
-          "local",
-          "export"
-        ]
-      }
-    ]
-  },
-  {
-    "type": "cycle",
-    "from": "src/lib/utils/index.ts",
-    "to": "src/lib/utils/narrativeParser.ts",
-    "rule": {
-      "severity": "warn",
-      "name": "no-circular"
-    },
-    "cycle": [
-      {
-        "name": "src/lib/utils/narrativeParser.ts",
-        "dependencyTypes": [
-          "local",
-          "export"
-        ]
-      },
-      {
-        "name": "src/lib/utils/index.ts",
-        "dependencyTypes": [
-          "local",
-          "import"
-        ]
-      }
-    ]
-  },
-  {
-    "type": "cycle",
-    "from": "src/lib/utils/index.ts",
-    "to": "src/lib/utils/textFormatter.ts",
-    "rule": {
-      "severity": "warn",
-      "name": "no-circular"
-    },
-    "cycle": [
-      {
-        "name": "src/lib/utils/textFormatter.ts",
-        "dependencyTypes": [
-          "local",
-          "export"
-        ]
-      },
-      {
-        "name": "src/lib/utils/index.ts",
-        "dependencyTypes": [
-          "aliased",
-          "aliased-tsconfig",
-          "aliased-tsconfig-paths",
-          "local",
-          "import"
-        ]
-      }
-    ]
-  },
-  {
-    "type": "cycle",
-    "from": "src/types/ai-testing.types.ts",
-    "to": "src/types/index.ts",
-    "rule": {
-      "severity": "warn",
-      "name": "no-circular"
-    },
-    "cycle": [
-      {
-        "name": "src/types/index.ts",
-        "dependencyTypes": [
-          "local",
-          "type-only",
-          "import"
-        ]
-      },
-      {
-        "name": "src/types/ai-testing.types.ts",
-        "dependencyTypes": [
-          "local",
-          "type-only",
-          "export"
-        ]
-      }
-    ]
   }
 ]
