+++ .skott-baseline.json
@@ -1,3 +1,3 @@
 {
-  "maxCircularDependencies": 6
+  "maxCircularDependencies": 1
 }
