+++ src/lib/theme/colorScheme.ts
@@ -1,0 +1,5 @@
+export type ColorScheme = 'light' | 'dark' | 'system';
+
+export const DEFAULT_COLOR_SCHEME: ColorScheme = 'light';
+
+export const STORAGE_KEY_COLOR_SCHEME = 'narraitor-color-scheme';
