+++ src/lib/utils/typeGuards.ts
@@ -6,3 +6,4 @@ export {
 } from './typeGuards/worldGuards';
 export { isPlayerDecisionArray } from './typeGuards/personalizationGuards';
 export { sanitizeString } from './typeGuards/stringGuards';
+export { isRecord, asTrimmedString, asArray } from './typeGuards/jsonGuards';
